+++ src/form/lean.cpp
@@ -11,6 +11,17 @@
 #include "sys/setup.hpp"
 #include "sys/util.hpp"
 
+std::string convertBitfuncToLean(const std::string& bitfunc) {
+  if (bitfunc == "bitand") {
+    return "Int.land";
+  } else if (bitfunc == "bitor") {
+    return "Int.lor";
+  } else if (bitfunc == "bitxor") {
+    return "Int.xor";
+  }
+  return "";
+}
+
 bool LeanFormula::convertToLean(Expression& expr, const Formula& f,
                                 const std::string& funcName) {
   // Check children recursively
@@ -67,12 +78,13 @@ bool LeanFormula::convertToLean(Expression& expr, const Formula& f,
         // If not a simple fraction, reject
         return false;
       }
-      // Convert bitxor to LEAN Int.xor
-      // if (expr.name == "bitxor") {
-      //  expr.name = "Int.xor";
-      //  imports.insert("import Mathlib.Data.Int.Bitwise");
-      //  break;
-      //}
+      // Convert bitwise functions
+      auto bitfunc = convertBitfuncToLean(expr.name);
+      if (!bitfunc.empty()) {
+        expr.name = bitfunc;
+        imports.insert("Mathlib.Data.Int.Bitwise");
+        break;
+      }
       // Allow recursive calls to the main function
       if (expr.name == funcName) {
         break;